# backend/spec/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/workspaces/{workspaceId}/export": {
            "get": {
                "operationId": "ExportController_exportTransactions",
                "parameters": [
                    {
                        "name": "format",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": ["csv", "json"]
                        }
                    },
                    {
                        "name": "date_from",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "format": "date",
                            "type": "string"
                        }
                    },
                    {
                        "name": "date_to",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "format": "date",
                            "type": "string"
                        }
                    },
                    {
                        "name": "workspaceId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "text/csv": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "tags": ["Export"]
            }
        },
...
    },
    "info": {
        "title": "Monqom API",
        "description": "The Monqom API description",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
...
}